test: ConfigurationBeanScanner 추가에 따른 코드 변경
