+++ nextstep-di/src/test/java/nextstep/di/factory/error/interfaces/BeanFactoryInterfaceErrorTest.java
@@ -1,12 +1,10 @@
 package nextstep.di.factory.error.interfaces;
 
-import nextstep.di.BeanScanner;
+import nextstep.di.AnnotatedTypeBeanScanner;
 import nextstep.di.factory.BeanFactory;
-import nextstep.di.factory.exception.IllegalAnnotationException;
+import nextstep.di.factory.exception.NoSuchDefaultConstructorException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-
-import java.util.Set;
 
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
@@ -18,9 +16,7 @@ public class BeanFactoryInterfaceErrorTest {
     @SuppressWarnings("unchecked")
     @DisplayName("인터페이스에 어노테이션을 달았을 때 예외 발생")
     void initialize() {
-        BeanScanner beanScanner = new BeanScanner("nextstep.di.factory.error.interfaces");
-        Set<Class<?>> preInstantiateClazz = beanScanner.scan();
-        beanFactory = new BeanFactory(preInstantiateClazz);
-        assertThrows(IllegalAnnotationException.class, () -> beanFactory.initialize());
+        AnnotatedTypeBeanScanner annotatedTypeBeanScanner = new AnnotatedTypeBeanScanner(new String[]{"nextstep.di.factory.error.interfaces"});
+        assertThrows(NoSuchDefaultConstructorException.class, annotatedTypeBeanScanner::scan);
     }
 }
