--- nextstep-di/src/test/java/nextstep/di/factory/example/ExampleConfig.java
@@ -1,20 +1,0 @@
-package nextstep.di.factory.example;
-
-import nextstep.annotation.Bean;
-import nextstep.annotation.Configuration;
-import org.apache.commons.dbcp2.BasicDataSource;
-
-import javax.sql.DataSource;
-
-@Configuration
-public class ExampleConfig {
-    @Bean
-    public DataSource dataSource() {
-        BasicDataSource ds = new BasicDataSource();
-        ds.setDriverClassName("org.h2.Driver");
-        ds.setUrl("jdbc:h2:~/jwp-framework;MVCC=TRUE;DB_CLOSE_ON_EXIT=FALSE");
-        ds.setUsername("sa");
-        ds.setPassword("");
-        return ds;
-    }
-}
