+++ nextstep-di/src/test/java/nextstep/di/factory/example/JdbcQuestionRepository.java
@@ -4,5 +4,6 @@ import nextstep.stereotype.Repository;
 
 @Repository
 public class JdbcQuestionRepository implements QuestionRepository {
-
+    public JdbcQuestionRepository() {
+    }
 }
